move notify-related code from `QueryInfo` to `ObservableQuery`
